create util function for copying directories
